% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/przygotowanie_danych.R
\name{.fuzzify_vector}
\alias{.fuzzify_vector}
\title{Internal Fuzzifier Function}
\usage{
.fuzzify_vector(vector)
}
\description{
Converts crisp number to Triangular Fuzzy Number (TFN).
TFN is a triple (l, m, u), where m = x, l = x-1, u = x+1.
}
\keyword{internal}
