% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiles.R
\name{.get_life_profile}
\alias{.get_life_profile}
\title{City Life Profiles}
\usage{
.get_life_profile(profile)
}
\description{
Predefiniowane profile uzytkownikow do analizy miast.
Kazdy profil zawiera wagi kryteriow i progi akceptowalnosci dopasowane
do potrzeb konkretnej grupy uzytkownikow.
}
\keyword{internal}
