% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/przygotowanie_danych.R
\name{.parse_mcda_syntax}
\alias{.parse_mcda_syntax}
\title{Internal MCDA Syntax Parser}
\usage{
.parse_mcda_syntax(syntax)
}
\description{
Helper function to interpret user-defined model syntax.
Converts text "Criterion =~ var1 + var2" into R list structure.
}
\keyword{internal}
