% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/przygotowanie_danych.R
\name{.scale_to_saaty}
\alias{.scale_to_saaty}
\title{Internal Saaty Scaler}
\usage{
.scale_to_saaty(vector)
}
\description{
Transforms any scale (e.g., Likert 1-5, continuous values)
to Saaty scale 1-9.
}
\keyword{internal}
