% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot.fuzzy_vikor_res}
\alias{plot.fuzzy_vikor_res}
\title{Plot Fuzzy VIKOR Results}
\usage{
\method{plot}{fuzzy_vikor_res}(x, ...)
}
\arguments{
\item{x}{Obiekt klasy fuzzy_vikor_res zwrocony przez funkcje fuzzy_vikor().}

\item{...}{Dodatkowe argumenty.}
}
\value{
Obiekt ggplot.
}
\description{
Wizualizuje wyniki metody Fuzzy VIKOR jako wykres babelkowy.
Os X reprezentuje uzytecznosc grupowa (S), os Y indywidualny zal (R).
Rozmiar babla jest odwrotnie proporcjonalny do wskaznika kompromisu Q -
im mniejsze Q (lepszy ranking), tym wiekszy babel.
}
