% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot.rozmyty_multimoora_wynik}
\alias{plot.rozmyty_multimoora_wynik}
\title{Mapa Strategiczna MULTIMOORA}
\usage{
\method{plot}{rozmyty_multimoora_wynik}(x, ...)
}
\description{
Mapa Strategiczna MULTIMOORA
}
