% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot.rozmyty_promethee_wynik}
\alias{plot.rozmyty_promethee_wynik}
\title{Wykres Przepływów PROMETHEE II}
\usage{
\method{plot}{rozmyty_promethee_wynik}(x, ...)
}
\description{
Wykres Przepływów PROMETHEE II
}
