% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy_mcdm.R
\name{rozmyty_multimoora}
\alias{rozmyty_multimoora}
\title{Rozmyta Metoda MULTIMOORA}
\usage{
rozmyty_multimoora(
  decision_mat,
  criteria_types,
  weights = NULL,
  bwm_criteria,
  bwm_best,
  bwm_worst
)
}
\arguments{
\item{decision_mat}{Matrix of TFN values (n x 3m format)}

\item{criteria_types}{Vector of "max"/"min" indicating benefit/cost criteria}

\item{weights}{Optional numeric vector of fuzzy weights (length = 3 * number of criteria)}

\item{bwm_criteria}{Optional criteria names for BWM weight calculation}

\item{bwm_best}{Optional Best-to-Others vector for BWM}

\item{bwm_worst}{Optional Others-to-Worst vector for BWM}
}
\value{
Obiekt klasy `rozmyty_multimoora_wynik`.
}
\description{
Implementacja metody Fuzzy MULTIMOORA. Składa się z:
1. Ratio System (RS)
2. Reference Point (RP)
3. Full Multiplicative Form (FMF)
Finalny ranking powstaje przez agregację Teorią Dominacji.
}
