% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rozmyte_promethee.R
\name{rozmyty_promethee}
\alias{rozmyty_promethee}
\title{Rozmyta Metoda PROMETHEE II}
\usage{
rozmyty_promethee(
  decision_mat,
  preference_params,
  weights = NULL,
  bwm_criteria,
  bwm_best,
  bwm_worst
)
}
\arguments{
\item{decision_mat}{Rozmyta macierz danych (n x 3m format)}

\item{preference_params}{Ramka danych z kolumnami: Type, q, p, s, Role ("min"/"max")}

\item{weights}{Wektor wag (opcjonalny)}

\item{bwm_criteria}{Optional criteria names for BWM weight calculation}

\item{bwm_best}{Optional Best-to-Others vector for BWM}

\item{bwm_worst}{Optional Others-to-Worst vector for BWM}
}
\value{
Obiekt klasy `rozmyty_promethee_wynik`.
}
\description{
Implementacja metody Fuzzy PROMETHEE. Oblicza przepływy netto (Phi).
}
