% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabela_apa.R
\name{tabela_apa}
\alias{tabela_apa}
\title{Generowanie Tabeli APA}
\usage{
tabela_apa(x, tytul = NULL)
}
\arguments{
\item{x}{Obiekt wynikowy z funkcji pakietu.}

\item{tytul}{Opcjonalny tytul tabeli.}
}
\description{
Funkcja przeksztalca wyniki analizy MCDA w tabele zgodna ze standardem APA.
}
