% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabela_apa.R
\name{zapisz_tabele}
\alias{zapisz_tabele}
\title{Zapisz tabele do Worda}
\usage{
zapisz_tabele(tabela, sciezka)
}
\arguments{
\item{tabela}{Obiekt flextable}

\item{sciezka}{Sciezka do pliku wyjsciowego}
}
\description{
Zapisuje tabele flextable do pliku .docx
}
