% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.calculate_entropy_weights}
\alias{.calculate_entropy_weights}
\title{Entropy-based Objective Weights (Fuzzy)}
\usage{
.calculate_entropy_weights(decision_mat)
}
\description{
Calculates objective fuzzy weights using Shannon entropy.
}
\keyword{internal}
