% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.get_final_weights}
\alias{.get_final_weights}
\title{Process Weights for Fuzzy MCDM}
\usage{
.get_final_weights(decision_mat, weights, bwm_criteria, bwm_best, bwm_worst)
}
\description{
Priority: explicit weights -> BWM -> entropy
}
\keyword{internal}
