% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy_vikor.R
\name{fuzzy_vikor}
\alias{fuzzy_vikor}
\title{Fuzzy VIKOR Method}
\usage{
fuzzy_vikor(
  decision_mat,
  criteria_types,
  v = 0.5,
  weights,
  bwm_criteria,
  bwm_best,
  bwm_worst
)
}
\arguments{
\item{decision_mat}{Matrix of TFN values (n x 3m format)}

\item{criteria_types}{Vector of 'max'/'min' indicating benefit/cost criteria}

\item{v}{Weight of the strategy of maximum group utility (default 0.5)}

\item{weights}{Optional numeric vector of fuzzy weights (length = 3 * number of criteria)}

\item{bwm_criteria}{Optional criteria names for BWM weight calculation}

\item{bwm_best}{Optional Best-to-Others vector for BWM}

\item{bwm_worst}{Optional Others-to-Worst vector for BWM}
}
\value{
Obiekt klasy fuzzy_vikor_res z wynikami rankingu.
}
\description{
Fuzzy VIKOR (VIseKriterijumska Optimizacija I Kompromisno Resenje).
Metoda koncentruje sie na wyborze rozwiazania kompromisowego, ktore maksymalizuje
uzytecznosc grupowa przy jednoczesnym minimalizowaniu indywidualnego zalu decydenta.
W pakiecie CityOptimaR pelni role metody weryfikujacej wyniki Fuzzy TOPSIS.
}
