% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot.fuzzy_vikor_res}
\alias{plot.fuzzy_vikor_res}
\title{Plot Fuzzy VIKOR Results}
\usage{
\method{plot}{fuzzy_vikor_res}(x, ...)
}
\arguments{
\item{x}{An object of class `fuzzy_vikor_res` returned by the `fuzzy_vikor()` function.}

\item{...}{Additional arguments.}
}
\value{
A ggplot object.
}
\description{
Visualizes the VIKOR results using a bubble chart.
The X-axis represents Group Utility (S), the Y-axis represents Individual Regret (R).
Bubble size is inversely proportional to the Compromise Solution (Q) - smaller Q (better rank) = larger bubble.
}
