% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabela_apa.R
\name{zapisz_tabele}
\alias{zapisz_tabele}
\title{Zapisz tabelę do Worda}
\usage{
zapisz_tabele(tabela, sciezka)
}
\arguments{
\item{tabela}{Obiekt flextable (wynik funkcji tabela_apa)}

\item{sciezka}{Ścieżka do pliku wyjściowego (np. "wyniki.docx")}
}
\description{
Pomocnicza funkcja do zapisu tabeli flextable do pliku .docx
}
